src/demo/requirements.txt
# Python requirements for MLOps Demo
# Core data processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.2.0
mlflow>=2.8.0

# Stream processing
apache-flink>=1.17.0

# Data storage and messaging
kafka-python>=2.0.2
boto3>=1.26.0
minio>=7.1.0

# Monitoring and observability
prometheus-client>=0.15.0
psutil>=5.9.0

# Utilities
requests>=2.28.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=0.19.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
